changed skills from About component to Resume component
